+++ CHANGES.rst
@@ -16,6 +16,12 @@ New features:
 
 - Update mockup to latest version.
   [thet]
+  
+- add registry settings for thumb and icon handling  in tables, lists and portlets
+  https://github.com/plone/Products.CMFPlone/issues/1734 (PLIP) 
+  recompiled bundle plone-logged-in
+  requires upgrade step (reapply profile)
+  [fgrcon]
 
 - new metadata catalog column mime_type
   https://github.com/plone/Products.CMFPlone/issues/1995